proper whitespace handling
this does not seem to conform to the best practice described for parslet
(i.e. put spaces to the lowest level and if possible to the end of the
rule).

However following the best practice would increase the complexity of the
intermediate tree even more... let's see if the approach used by me
scales...